Split up the monolithic Core module
