+++ Language/Haskell/TH/Desugar/Match.hs
@@ -32,6 +32,7 @@ import qualified Data.Map as Map
 import Language.Haskell.TH.Instances ()
 import Language.Haskell.TH.Syntax
 
+import Language.Haskell.TH.Desugar.AST
 import Language.Haskell.TH.Desugar.Core
 import Language.Haskell.TH.Desugar.Util
 import Language.Haskell.TH.Desugar.Reify
