+++ Language/Haskell/TH/Desugar/Subst.hs
@@ -29,6 +29,7 @@ import qualified Data.Set as S
 import Data.Generics
 import Data.List
 
+import Language.Haskell.TH.Desugar.AST
 import Language.Haskell.TH.Desugar.Core
 import Language.Haskell.TH.Syntax
 import Language.Haskell.TH.Desugar.Util
