+++ th-desugar.cabal
@@ -55,7 +55,8 @@ library
                       Language.Haskell.TH.Desugar.Lift,
                       Language.Haskell.TH.Desugar.Expand,
                       Language.Haskell.TH.Desugar.Subst
-  other-modules:      Language.Haskell.TH.Desugar.Core,
+  other-modules:      Language.Haskell.TH.Desugar.AST,
+                      Language.Haskell.TH.Desugar.Core,
                       Language.Haskell.TH.Desugar.Match,
                       Language.Haskell.TH.Desugar.Util,
                       Language.Haskell.TH.Desugar.Reify
